Restructure conformance folder structure.
