+++ conformance/src/js/fastify/app.js
@@ -1,4 +1,4 @@
-import { ConformanceApiService } from "../../../ts/src/conformanceApiService.js";
+import { ConformanceApiService } from "../../ts/conformanceApiService.js";
 import conformanceTestsJson from "../../../ConformanceTests.json";
 import { conformanceApiPlugin } from "./conformanceApiPlugin.js";
