+++ conformance/test/conformanceApiTests.ts
@@ -1,21 +1,21 @@
-import { createHttpClient } from '../src/conformanceApi';
-import { expect, should } from 'chai';
-import fetch from 'node-fetch';
-import conformanceTestsJson from '../../ConformanceTests.json';
+import { createHttpClient } from "../src/ts/conformanceApi";
+import { expect, should } from "chai";
+import fetch from "node-fetch";
+import conformanceTestsJson from "../ConformanceTests.json";
 
 should();
 
 const httpClient = createHttpClient({
 	fetch: (uri, request) => {
-		return fetch('http://localhost:4117/' + uri, request);
-	}
+		return fetch("http://localhost:4117/" + uri, request);
+	},
 });
 
-describe('tests', () => {
-
+describe("tests", () => {
 	conformanceTestsJson.tests.forEach((data: any) => {
 		it(data.test, async () => {
-			return ((httpClient as any)[data.method](data.request))
+			return (httpClient as any)
+				[data.method](data.request)
 				.then((result: any) => {
 					expect({
 						error: result.error ?? undefined,
@@ -27,5 +27,4 @@ describe('tests', () => {
 				});
 		});
 	});
-
 });
