+++ conformance/tsconfig.json
@@ -13,7 +13,8 @@
     "allowJs": true
   },
   "include": [
-    "ts/**/*.ts",
-    "js/**/*.js"
-  ]
+    "src/**/*.ts",
+    "src/**/*.js",
+    "test/**/*.ts"
+  ],
 }
